Project structure overhaul while adding more set logic foundation
