--- c/collections.h
@@ -1,12 +1,0 @@
-#ifndef dictu_collections_h
-#define dictu_collections_h
-
-#include <stdbool.h>
-#include <string.h>
-#include <stdlib.h>
-
-bool listMethods(char *method, int argCount);
-
-bool dictMethods(char *method, int argCount);
-
-#endif //dictu_collections_h
