+++ c/datatypes/dicts.h
@@ -1,0 +1,12 @@
+#ifndef dictu_dicts_h
+#define dictu_dicts_h
+
+#include <stdbool.h>
+#include <string.h>
+#include <stdlib.h>
+
+#include "copy.h"
+
+bool dictMethods(char *method, int argCount);
+
+#endif //dictu_dicts_h
