+++ c/datatypes/lists.h
@@ -1,0 +1,12 @@
+#ifndef dictu_lists_h
+#define dictu_lists_h
+
+#include <stdbool.h>
+#include <string.h>
+#include <stdlib.h>
+
+#include "copy.h"
+
+bool listMethods(char *method, int argCount);
+
+#endif //dictu_lists_h
