+++ c/datatypes/sets.h
@@ -1,0 +1,12 @@
+#ifndef dictu_sets_h
+#define dictu_sets_h
+
+#include <stdbool.h>
+#include <string.h>
+#include <stdlib.h>
+
+#include "copy.h"
+
+bool setMethods(char *method, int argCount);
+
+#endif //dictu_sets_h
