+++ c/datatypes/strings.h
